Fix dtype in ufunc_group_by_idx

+++ xminds/lib/aggregate.py
@@ -153,7 +153,7 @@ def ufunc_group_by_idx(idx, values, ufunc, init, minlength=None):
     array([ 10, 12, -0,  10])
     """
     length = max(idx.max() + 1 if idx.size else 0, minlength or 0)
-    out = numpy.full(length, init)
+    out = numpy.full(length, init, dtype=values.dtype)
     ufunc.at(out, idx, values)
     return out
